improved speed of folding
git-svn-id: file:///usr/local/SVN/trunk@218 f794847c-437f-4cdd-b6dc-fa27500bcbe7

+++ MolGenExp/src/biochem/AcidInChain.java
@@ -91,11 +91,11 @@ public class AcidInChain implements Serializable {
 		return a.getHydrophobicIndex();
 	}
 
-	public double gethydrogenbondIndex() {
+	public int gethydrogenbondIndex() {
 		return a.gethydrogenbondIndex();
 	}
 
-	public double getionicIndex() {
+	public int getionicIndex() {
 		return a.getionicIndex();
 	}
